fixed compile utilities.

+++ app/utils/compile.php
@@ -11,23 +11,31 @@ use Symfony\Component\Console\Output\NullOutput;
 $vendor_dir = dirname(dirname(__DIR__)).'/vendor';
 require_once( $vendor_dir.'/autoload.php');
 
+#
+# set up class preLoader script.
+#
+$config = call_user_func(include(dirname(__DIR__).'/config.php'), ['debug' => false, 'xhProf' => false]);
 
-/*
- * set up class preLoader script.
- */
 $getIncludes = __DIR__.'/compile-includes.php';
-$outputPath  = dirname(dirname(__DIR__)).'/var/compiled.php';
+$outputPath  = $config['vars_dir'].'/compiled.php';
 
-/*
- * run compilation
- */
-unlink($outputPath);
+if (file_exists($outputPath)) {
+    unlink($outputPath);
+}
+if (file_exists($config['vars_dir'].'/app.cache')) {
+    unlink($config['vars_dir'].'/app.cache');
+}
+
+#
+# run compilation
+#
+
 $command = new PreCompileCommand();
 $command->run(
     new ArrayInput([
-            '--config' => $getIncludes,
-            '--output' => $outputPath,
-            '--strip_comments' => 1,
-        ]),
+        '--config' => $getIncludes,
+        '--output' => $outputPath,
+        '--strip_comments' => 1,
+    ]),
     new NullOutput
 );
